+++ compiler-core/src/type_/expression.rs
@@ -1273,6 +1273,15 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             }
             None => None,
         };
+
+        // We're checking if we have a prepend but no elements, so we can
+        // provide an error telling the developer to just use the list directly.
+        // Like this: `[..tail]`.
+        if inferred_elements.is_empty() && tail.is_some() {
+            self.problems
+                .error(Error::ListPrependWithoutElements { location })
+        }
+
         TypedExpr::List {
             location,
             type_,
@@ -4679,6 +4688,14 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         } else {
             None
         };
+
+        // We're checking if we have a prepend but no elements, so we can
+        // provide an error telling the developer to just use the list directly.
+        // Like this: `[..tail]`.
+        if elements.is_empty() && tail.is_some() {
+            self.problems
+                .error(Error::ListPrependWithoutElements { location })
+        }
 
         Constant::List {
             elements,