+++ CHANGELOG.md
@@ -28,6 +28,10 @@
   to optimise.
   ([John Downey](https://github.com/jtdowney))
 
+- The compiler is now fault tolerant when providing an error for a redundant list
+  prepend like `[..wobble]`.
+  ([0xda157](https://github.com/0xda157))
+
 ### Build tool
 
 - The build tool now stores its build cache in a more compact binary format,